Add command to show images from current page.
Summary of changes:

- Added three functions which implement the image display:
  `showImages`, `fetchImageTargets`, `fetchImageUrls`.
- Added :image command which calls `showImages`.
- Modified _query to accept a custom URL kwarg.
  (necessary for `fetchImageTargets`.)

Description:

The new :image command calls a program with the fetched images as a list
of space separated URLs. By default this is `feh` as it is a lightweight
and nifty image viewer.

As of this commit, this feature only works on Wikipedia. I may try to
extend that, though it may prove difficult as the API is rather lacking
in this department, and different wikis seems to use images differently.
I'll need to do some more research.

The program and arguments (flags etc) can be specified in the config file.
For example:

[images]
program = ristretto
arguments = foo, bar, baz

This will lead to the following being called upon execution:

subprocess.Popen(['ristretto', 'foo', 'bar', 'baz', 'url1', 'url2',...])

TODO:

More testing. Try to improve coverage so that we get more relevant
images. Implement some kind of error reporting as there is literally
zero as of this commit.